binutils: add support for dumping BTF info to objdump
This patch adds a new command line option --btf[=SECTION] to objdump
to dump a textual representation the BTF data stored in section
SECTION.  If no section is specified then the default .BTF is used.

The textual representation is the same than the one emitted by the
Linux kernel bpftool.

binutils/ChangeLog:

2023-11-23  Jose E. Marchesi  <jose.marchesi@oracle.com>

	* objdump.c: Include btf.h.
	(usage): Document --btf.
	(enum option_values): Add OPTION_BTF.
	(long_options): Entry for --btf.
	(dump_btf): New function.
	(main): Handle --btf.
	(dump_bfd): Call dump_btf_section whenever appropriate.
	* doc/btf.options.texi: New file.
	* doc/binutils.texi (objdump): Document --btf.